Изменил зависимости в ПОМнике
