--- HomeWork/src/Parser/Parser.java
@@ -1,6 +1,0 @@
-package Parser;
-
-
-public class Parser {
-
-}
